Move report table to own component
 * Report table is now it's own component
 * Activity reports table replaced with new stand alone component
 * My alerts still needs to be updated to use new component. There are a
few difference between the two tables that need to be handled. Saving
for a later date to keep the change smaller
 * withWidgetData uses deep compare to prevent multiple renders
 * Moved table related tests from landing page to new component
 * Add report table to grantee dashboard
