+++ frontend/src/widgets/withWidgetData.js
@@ -1,5 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
+import useDeepCompareEffect from 'use-deep-compare-effect';
 import PropTypes from 'prop-types';
 import fetchWidget from '../fetchers/Widgets';
 import { filtersToQueryString } from '../components/Filter';
@@ -18,7 +19,7 @@ const withWidgetData = (Widget, widgetId) => {
 
     const { filters } = props;
 
-    useEffect(() => {
+    useDeepCompareEffect(() => {
       const fetch = async () => {
         try {
           updateLoading(true);
